\name{+}
\alias{-,ANY,Reference-method}
\alias{-,Reference,ANY-method}
\alias{*,ANY,Reference-method}
\alias{*,Reference,ANY-method}
\alias{/,ANY,Reference-method}
\alias{/,Reference,ANY-method}
\alias{\%*\%,ANY,Reference-method}
\alias{\%*\%,Reference,ANY-method}
\alias{\%\%,ANY,Reference-method}
\alias{\%\%,Reference,ANY-method}
\alias{+}
\alias{+,ANY,Reference-method}
\alias{+,Reference,ANY-method}
\title{Perform arithmetic operations with objects held by a Reference class
instance.}
\description{
  Perform arithmetic operations with objects held by a
  Reference class instance.
}

