\name{crossprod}
\alias{crossprod}
\alias{crossprod,ANY,Reference-method}
\alias{crossprod,Reference,ANY-method}
\title{Get the matrix cross product when one of the matrices is maintained
by a Reference class instance.}
\description{
  Get the matrix cross product when one of the matrices is
  maintained by a Reference class instance.
}

