\docType{methods}
\name{[}
\alias{[}
\alias{[,Reference,ANY-logical-method}
\alias{[,Reference,ANY-method}
\alias{[,Reference,ANY,ANY-method}
\alias{[,Reference,ANY,character-method}
\alias{[,Reference,ANY,character-method,}
\alias{[,Reference,ANY,logical-method}
\alias{[,Reference,ANY,numeric-method}
\alias{[[,Reference,ANY,ANY-method}
\alias{[[,Reference,ANY,character-method}
\alias{[[,Reference,ANY,logical-method}
\alias{[[,Reference,ANY,numeric-method}
\alias{[[<-,Reference,ANY,ANY-method}
\alias{[[<-,Reference,ANY,character-method}
\alias{[[<-,Reference,ANY,logical-method}
\alias{[[<-,Reference,ANY,numeric-method}
\alias{[<-,Reference,ANY,ANY-method}
\alias{[<-,Reference,ANY,character-method}
\alias{[<-,Reference,ANY,logical-method}
\alias{[<-,Reference,ANY,numeric-method}
\title{Extract or replace values from a Reference object}
\description{
  These functions extract values from Refernce objects
  holding n-dimensional data structures.
}

